Introduce :launchpad/jvm-opts
Make sure we populate the internal :java-args from the above-mentioned
key, therefore allowing `deps.local.edn` to contain things that differ
from `deps.edn`.

Note that there is naming discrepancy as this patch sticks with the
official key - `:jvm-opts` rather than the internal name.

The behavior of appending to existing `:jvm-opts`, in order, has been
preserved as per the `tools.deps` built-in.